avoid subtraction wrap around

+++ src/util/rmalloc.c
@@ -44,7 +44,7 @@ static inline void _nmalloc_increment(int64_t n_bytes) {
 	if(n_alloced > mem_capacity) {
 		// set n_alloced to MIN to avoid further out of memory exceptions
 		// TODO: consider switching to double -inf
-		n_alloced = INT64_MIN;
+		n_alloced = INT32_MIN;
 
 		// throw exception cause memory limit exceeded
 		ErrorCtx_SetError("Query's mem consumption exceeded capacity");
